Moved some components
closes #322

Also:
- renamed `TimeLineHistogram.svelte` to `Timeline.svelte`
- renamed ViewSelector.svelte` to `ViewSelectorMedium.svelte` (used also in a small component but this will change soon)
- changed the main class in `StackedBarchart.svelte` to `.StackedBarchart`
- moved unused components in `unused` dirs
- formatting double spaces to single space or tabs
