+++ fe/src/lib/components/explorer/SelectionXor.svelte
