+++ fe/src/lib/components/explorer/SelectorInterval.svelte
@@ -1,5 +1,5 @@
 <script>
-	import SelectionXor from '$lib/components/explorer/medium/SelectionXor.svelte';
+	import SelectionXor from '$lib/components/explorer/SelectionXor.svelte';
 	import {intervalToLabel} from '$lib/config/labels.js';
 </script>
