+++ fe/src/lib/components/explorer/SelectorRegionType.svelte
@@ -1,7 +1,7 @@
 <script>
 	import * as _ from 'lamb';
 
-	import SelectionXor from '$lib/components/explorer/medium/SelectionXor.svelte';
+	import SelectionXor from '$lib/components/explorer/SelectionXor.svelte';
 	import {regionTypeToLabel} from '$lib/config/labels.js';
 	import {_selection} from '$lib/stores/navigation.js';
