+++ fe/src/lib/components/explorer/medium/ViewSelectorMedium.svelte
