+++ fe/src/lib/components/explorer/small/IconBar.svelte
