+++ fe/src/lib/components/svizzle/trends/StreamGraph.svelte
@@ -195,7 +195,7 @@
 
 			xScale = _.pipe([keyRankFn, timeScale]);
 		} else {
-			keyTicks =  _.map(
+			keyTicks = _.map(
 				keyFilterFn ? _.filter(allKeys, keyFilterFn) : allKeys,
 				_.collect([_.identity, keyFormatFn])
 			);
