+++ fe/src/lib/components/svizzle/trends/Trends.svelte
@@ -160,7 +160,7 @@
 
 			xScale = _.pipe([keyRankFn, timeScale]);
 		} else {
-			keyTicks =  _.map(
+			keyTicks = _.map(
 				keyFilterFn ? _.filter(allKeys, keyFilterFn) : allKeys,
 				_.collect([_.identity, keyFormatFn])
 			);
