+++ fe/src/routes/explorer/number/stats/[slug]/+page.svelte
@@ -8,16 +8,16 @@
 	import * as _ from 'lamb';
 
 	import {page as _page} from '$app/stores';
-    import MetricTitle from '$lib/components/explorer/MetricTitle.svelte';
+	import MetricTitle from '$lib/components/explorer/MetricTitle.svelte';
 	import Grid2Columns from '$lib/components/svizzle/Grid2Columns.svelte';
-    import GridRows from '$lib/components/svizzle/GridRows.svelte';
+	import GridRows from '$lib/components/svizzle/GridRows.svelte';
 	import Treemap from '$lib/components/svizzle/Treemap.svelte';
 	import View from '$lib/components/viewports/View.svelte';
+	import {_isSmallScreen} from '$lib/stores/layout.js';
 	import {_currentMetric} from '$lib/stores/navigation.js';
 	import {_isViewReady, _viewData} from '$lib/stores/view.js';
 	import {_histogramsTheme} from '$lib/stores/theme.js';
 	import {getDocCount} from '$lib/utils/getters.js';
-    import { _isSmallScreen } from '$lib/stores/layout.js';
 
 	const valueAccessor = getDocCount;
