+++ fe/src/routes/explorer/number/time/[slug]/+page.svelte
@@ -9,17 +9,15 @@
 	import {page as _page} from '$app/stores';
 	import FlexBar from '$lib/components/explorer/FlexBar.svelte';
 	import PercentilesTrendsView
-		from '$lib/components/explorer/medium/PercentilesTrendsView.svelte';
-	import SelectionXor
-		from '$lib/components/explorer/medium/SelectionXor.svelte';
-	import SelectorInterval
-		from '$lib/components/explorer/medium/SelectorInterval.svelte';
+		from '$lib/components/explorer/PercentilesTrendsView.svelte';
+	import SelectionXor from '$lib/components/explorer/SelectionXor.svelte';
+	import SelectorInterval from '$lib/components/explorer/SelectorInterval.svelte';
 	import MetricTitle from '$lib/components/explorer/MetricTitle.svelte';
 	import Grid2Rows from '$lib/components/svizzle/Grid2Rows.svelte';
-    import GridRows from '$lib/components/svizzle/GridRows.svelte';
+	import GridRows from '$lib/components/svizzle/GridRows.svelte';
 	import Trends from '$lib/components/svizzle/trends/Trends.svelte';
 	import View from '$lib/components/viewports/View.svelte';
-    import {_isSmallScreen} from '$lib/stores/layout.js';
+	import {_isSmallScreen} from '$lib/stores/layout.js';
 	import {_currentMetric, _selection} from '$lib/stores/navigation.js';
 	import {_currThemeVars, _framesTheme} from '$lib/stores/theme.js';
 	import {_isViewReady, _viewData} from '$lib/stores/view.js';
